Rename generate_preSignedUrl.js to s3_operator.js Add logic that returning BucketName
